swiftvalidator/src/net/uglobal/swiftvalidator/swift/field/Field90K.java

package net.uglobal.swiftvalidator.swift.field;



/**
 * Field format : (qualifier)(indexPoints)
 * 
 * AUTO GENERATED - DO NOT HAND MODIFY
 * Generated on Sun Dec 09 17:22:28 EST 2012
 * 
 */
public class Field90K
    extends net.uglobal.swiftvalidator.swift.field.Field
{


    public Field90K(String validateData, net.uglobal.swiftvalidator.xml.beans.Field xmlField) {
        super(validateData, xmlField);
    }

    @Override
    protected void parse() {
        addComponents(net.uglobal.swiftvalidator.swift.field.FieldComponentParseUtil.parse(validateData, getContentFormat()));
    }

    public String getQualifier() {
        return getComponent1();
    }

    public String getIndexPoints() {
        return getComponent2();
    }

    public String toString() {
        return (new StringBuilder()).append("[" ).append( "qualifier=" + getComponent1()).append( ", indexPoints=" + getComponent2() ).append( "]").toString();
    }

}
